\name{Liszt-package}
\alias{Liszt}
\alias{ListLength}
\alias{List Length Analysis}
\alias{List Length Analysis-package}
\docType{package}
\title{
List Length Analysis
}
\description{
Package to perform list length analysis using an interface to the \code{jags} software.
}
\details{
\tabular{ll}{
Package: \tab ListLength\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-10\cr
License: \tab GPL (>= 2)\cr
}
Allows the user to perform list length analysis, as described in the
reference below, using R and R2jags...
}
\author{ 
	Megan Barnes <megan.barnes@uq.edu.au>
}
\references{
  Barnes, M and Szabo, JK. Guide to List Length Analysis and the Liszt package. 
  Available Online at: \url{http://www.edg.org.au/free-tools/listlength.html}
  
  Szabo JK, Vesk PA, Baxter PWJ, Possingham HP (2010) Regional avian
  species declines estimated from volunteer-collected long-term data
  using List Length Analysis. Ecological Applications 20
  (8):2157-2169. doi:10.1890/09-0877.1
  
  Barnes, M, Szabo JK, Possingham HP (In prep) No difference in trends 
  of endemic avifauna within and outside of protected areas. 

  Further guidance on using this package and associated functions and list length analysis in general
  is available at: \url{http://www.edg.org.au/free-tools/listlength.html}
}
\keyword{ package }
\seealso{
\code{\link[rjags:rjags-package]{rjags}}
\code{\link[R2jags]{jags}}
\code{A3Model}
\code{A5Model}
}
\examples{
### See functions \code{A3Model} and \code{A5Model}.
}
